backend/requirements.txt
fastapi
pydantic
uvicorn[standard]
google-genai
eth-account
python-multipart
pillow
slowapi
google-cloud-secret-manager
pytest
pytest-asyncio
httpx
web3
python-dotenv
numpy
sqlalchemy
psycopg2-binary
